+++ src/main/java/com/team3256/database/model/inventory/PartVendorInformationRepository.java
@@ -2,5 +2,5 @@ package com.team3256.database.model.inventory;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface PartVendorInformationRepository extends CrudRepository<PartVendorInformation, Long> {
+public interface PartVendorInformationRepository extends CrudRepository<PartVendorInformation, Integer> {
 }
